refactor(api): move CalibratedChallenge and CalibratedChallengeSkill builders to certif folder
