+++ api/tests/tooling/domain-builder/factory/certification/evaluation/build-calibrated-challenge-skill.js
@@ -1,4 +1,4 @@
-import { CalibratedChallengeSkill } from '../../../../src/certification/evaluation/domain/models/CalibratedChallengeSkill.js';
+import { CalibratedChallengeSkill } from '../../../../../../src/certification/evaluation/domain/models/CalibratedChallengeSkill.js';
 
 export function buildCalibratedChallengeSkill({
   id = 'recSkill1',
